requirements.txt
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
flake8==3.8.4
idna==2.10
mccabe==0.6.1
numpy==1.20.0
pandas==1.2.1
pycodestyle==2.6.0
pyflakes==2.2.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
rope==0.18.0
six==1.15.0
toml==0.10.2
urllib3==1.26.3
